AI-Medical-Scheduler/AI-patient-scheduler/requirements.txt
streamlit==1.37.1
pandas==2.2.2
openpyxl==3.1.5
xlsxwriter==3.2.0
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.8.2
langchain==0.2.12
langgraph==0.2.35
ollama==0.3.1
openai==1.45.0  # optional fallback
httpx==0.27.0
uvicorn==0.30.6
fastapi==0.114.0  # optional, only if you add APIs
tenacity==8.5.0
twilio==9.2.3   # optional, SMS
